requirements.txt
asgiref==3.2.10
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
dj-database-url==0.5.0
Django==3.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.4
typing-extensions==4.6.3
urllib3==1.26.16
whitenoise==6.5.0
